@@ -1,12 +1,10 @@
 name: flutter_html
 description: A Flutter widget rendering static HTML and CSS as Flutter widgets.
-version: 2.2.4
-homepage: https://github.com/Sub6Resources/flutter_html
+version: 3.0.0
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.2.0"
+  sdk: ^3.4.0
 
 dependencies:
   # Plugin for parsing html
