@@ -13,15 +13,15 @@ import 'package:html/parser.dart' as htmlparser;
 import 'package:numerus/numerus.dart';
 
 typedef OnTap = void Function(
-    String? url,
-    RenderContext context,
-    Map<String, String> attributes,
-    dom.Element? element,
+  String? url,
+  RenderContext context,
+  Map<String, String> attributes,
+  dom.Element? element,
 );
 typedef OnMathError = Widget Function(
-    String parsedTex,
-    String exception,
-    String exceptionWithType,
+  String parsedTex,
+  String exception,
+  String exceptionWithType,
 );
 typedef OnCssParseError = String? Function(
   String css,
@@ -72,15 +72,17 @@ class HtmlParser extends StatelessWidget {
     this.selectionControls,
     this.scrollPhysics,
   })  : this.internalOnAnchorTap = onAnchorTap != null
-          ? onAnchorTap
-          : key != null
-              ? _handleAnchorTap(key, onLinkTap)
-              : null,
+            ? onAnchorTap
+            : key != null
+                ? _handleAnchorTap(key, onLinkTap)
+                : null,
         super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    Map<String, Map<String, List<css.Expression>>> declarations = _getExternalCssDeclarations(htmlData.getElementsByTagName("style"), onCssParseError);
+    Map<String, Map<String, List<css.Expression>>> declarations =
+        _getExternalCssDeclarations(
+            htmlData.getElementsByTagName("style"), onCssParseError);
     StyledElement lexedTree = lexDomTree(
       htmlData,
       customRenders.keys.toList(),
@@ -93,8 +95,10 @@ class HtmlParser extends StatelessWidget {
     if (declarations.isNotEmpty) {
       externalCssStyledTree = _applyExternalCss(declarations, lexedTree);
     }
-    StyledElement inlineStyledTree = _applyInlineStyles(externalCssStyledTree ?? lexedTree, onCssParseError);
-    StyledElement customStyledTree = _applyCustomStyles(style, inlineStyledTree);
+    StyledElement inlineStyledTree =
+        _applyInlineStyles(externalCssStyledTree ?? lexedTree, onCssParseError);
+    StyledElement customStyledTree =
+        _applyCustomStyles(style, inlineStyledTree);
     StyledElement cascadedStyledTree = _cascadeStyles(style, customStyledTree);
     StyledElement cleanedTree = cleanTree(cascadedStyledTree);
     InlineSpan parsedTree = parseTree(
@@ -162,7 +166,7 @@ class HtmlParser extends StatelessWidget {
       name: "[Tree Root]",
       children: <StyledElement>[],
       node: html.documentElement,
-      style: Style.fromTextStyle(Theme.of(context).textTheme.bodyText2!),
+      style: Style.fromTextStyle(Theme.of(context).textTheme.bodyMedium!),
     );
 
     html.nodes.forEach((node) {
@@ -214,7 +218,8 @@ class HtmlParser extends StatelessWidget {
       } else if (INTERACTABLE_ELEMENTS.contains(node.localName)) {
         return parseInteractableElement(node, children);
       } else if (REPLACED_ELEMENTS.contains(node.localName)) {
-        return parseReplacedElement(node, children, navigationDelegateForIframe);
+        return parseReplacedElement(
+            node, children, navigationDelegateForIframe);
       } else if (LAYOUT_ELEMENTS.contains(node.localName)) {
         return parseLayoutElement(node, children);
       } else if (TABLE_CELL_ELEMENTS.contains(node.localName)) {
@@ -225,26 +230,30 @@ class HtmlParser extends StatelessWidget {
         final StyledElement tree = parseStyledElement(node, children);
         for (final entry in customRenderMatchers) {
           if (entry.call(
-              RenderContext(
-                buildContext: context,
-                parser: parser,
-                tree: tree,
-                style: Style.fromTextStyle(Theme.of(context).textTheme.bodyText2!),
-              ),
-            )) {
+            RenderContext(
+              buildContext: context,
+              parser: parser,
+              tree: tree,
+              style:
+                  Style.fromTextStyle(Theme.of(context).textTheme.bodyMedium!),
+            ),
+          )) {
             return tree;
           }
         }
         return EmptyContentElement();
       }
     } else if (node is dom.Text) {
-      return TextContentElement(text: node.text, style: Style(), element: node.parent, node: node);
+      return TextContentElement(
+          text: node.text, style: Style(), element: node.parent, node: node);
     } else {
       return EmptyContentElement();
     }
   }
 
-  static Map<String, Map<String, List<css.Expression>>> _getExternalCssDeclarations(List<dom.Element> styles, OnCssParseError? errorHandler) {
+  static Map<String, Map<String, List<css.Expression>>>
+      _getExternalCssDeclarations(
+          List<dom.Element> styles, OnCssParseError? errorHandler) {
     String fullCss = "";
     for (final e in styles) {
       fullCss = fullCss + e.innerHtml;
@@ -257,7 +266,9 @@ class HtmlParser extends StatelessWidget {
     }
   }
 
-  static StyledElement _applyExternalCss(Map<String, Map<String, List<css.Expression>>> declarations, StyledElement tree) {
+  static StyledElement _applyExternalCss(
+      Map<String, Map<String, List<css.Expression>>> declarations,
+      StyledElement tree) {
     declarations.forEach((key, style) {
       try {
         if (tree.matchesSelector(key)) {
@@ -271,7 +282,8 @@ class HtmlParser extends StatelessWidget {
     return tree;
   }
 
-  static StyledElement _applyInlineStyles(StyledElement tree, OnCssParseError? errorHandler) {
+  static StyledElement _applyInlineStyles(
+      StyledElement tree, OnCssParseError? errorHandler) {
     if (tree.attributes.containsKey("style")) {
       final newStyle = inlineCssToStyle(tree.attributes['style'], errorHandler);
       if (newStyle != null) {
@@ -285,7 +297,8 @@ class HtmlParser extends StatelessWidget {
 
   /// [applyCustomStyles] applies the [Style] objects passed into the [Html]
   /// widget onto the [StyledElement] tree, no cascading of styles is done at this point.
-  static StyledElement _applyCustomStyles(Map<String, Style> style, StyledElement tree) {
+  static StyledElement _applyCustomStyles(
+      Map<String, Style> style, StyledElement tree) {
     style.forEach((key, style) {
       try {
         if (tree.matchesSelector(key)) {
@@ -300,7 +313,8 @@ class HtmlParser extends StatelessWidget {
 
   /// [_cascadeStyles] cascades all of the inherited styles down the tree, applying them to each
   /// child that doesn't specify a different style.
-  static StyledElement _cascadeStyles(Map<String, Style> style, StyledElement tree) {
+  static StyledElement _cascadeStyles(
+      Map<String, Style> style, StyledElement tree) {
     tree.children.forEach((child) {
       child.style = tree.style.copyOnlyInherited(child.style);
       _cascadeStyles(style, child);
@@ -340,23 +354,34 @@ class HtmlParser extends StatelessWidget {
 
     for (final entry in customRenders.keys) {
       if (entry.call(newContext)) {
-        final buildChildren = () => tree.children.map((tree) => parseTree(newContext, tree)).toList();
-        if (newContext.parser.selectable && customRenders[entry] is SelectableCustomRender) {
-          final selectableBuildChildren = () => tree.children.map((tree) => parseTree(newContext, tree) as TextSpan).toList();
-          return (customRenders[entry] as SelectableCustomRender).textSpan.call(newContext, selectableBuildChildren);
+        final buildChildren = () =>
+            tree.children.map((tree) => parseTree(newContext, tree)).toList();
+        if (newContext.parser.selectable &&
+            customRenders[entry] is SelectableCustomRender) {
+          final selectableBuildChildren = () => tree.children
+              .map((tree) => parseTree(newContext, tree) as TextSpan)
+              .toList();
+          return (customRenders[entry] as SelectableCustomRender)
+              .textSpan
+              .call(newContext, selectableBuildChildren);
         }
         if (newContext.parser.selectable) {
-          return customRenders[entry]!.inlineSpan!.call(newContext, buildChildren) as TextSpan;
+          return customRenders[entry]!
+              .inlineSpan!
+              .call(newContext, buildChildren) as TextSpan;
         }
         if (customRenders[entry]?.inlineSpan != null) {
-          return customRenders[entry]!.inlineSpan!.call(newContext, buildChildren);
+          return customRenders[entry]!
+              .inlineSpan!
+              .call(newContext, buildChildren);
         }
         return WidgetSpan(
           child: ContainerSpan(
             newContext: newContext,
             style: tree.style,
             shrinkWrap: newContext.parser.shrinkWrap,
-            child: customRenders[entry]!.widget!.call(newContext, buildChildren),
+            child:
+                customRenders[entry]!.widget!.call(newContext, buildChildren),
           ),
         );
       }
@@ -364,10 +389,13 @@ class HtmlParser extends StatelessWidget {
     return WidgetSpan(child: Container(height: 0, width: 0));
   }
 
-  static OnTap _handleAnchorTap(Key key, OnTap? onLinkTap) =>
-          (String? url, RenderContext context, Map<String, String> attributes, dom.Element? element) {
+  static OnTap _handleAnchorTap(Key key, OnTap? onLinkTap) => (String? url,
+          RenderContext context,
+          Map<String, String> attributes,
+          dom.Element? element) {
         if (url?.startsWith("#") == true) {
-          final anchorContext = AnchorKey.forId(key, url!.substring(1))?.currentContext;
+          final anchorContext =
+              AnchorKey.forId(key, url!.substring(1))?.currentContext;
           if (anchorContext != null) {
             Scrollable.ensureVisible(anchorContext);
           }
@@ -411,24 +439,33 @@ class HtmlParser extends StatelessWidget {
       /// initialize indices to negative numbers to make conditionals a little easier
       int textIndex = -1;
       int elementIndex = -1;
+
       /// initialize parent after to a whitespace to account for elements that are
       /// the last child in the list of elements
       String parentAfterText = " ";
+
       /// find the index of the text in the current tree
       if ((tree.element?.nodes.length ?? 0) >= 1) {
-        textIndex = tree.element?.nodes.indexWhere((element) => element == tree.node) ?? -1;
+        textIndex =
+            tree.element?.nodes.indexWhere((element) => element == tree.node) ??
+                -1;
       }
+
       /// get the parent nodes
       dom.NodeList? parentNodes = tree.element?.parent?.nodes;
+
       /// find the index of the tree itself in the parent nodes
       if ((parentNodes?.length ?? 0) >= 1) {
-        elementIndex = parentNodes?.indexWhere((element) => element == tree.element) ?? -1;
+        elementIndex =
+            parentNodes?.indexWhere((element) => element == tree.element) ?? -1;
       }
+
       /// if the tree is any node except the last node in the node list and the
       /// next node in the node list is a text node, then get its text. Otherwise
       /// the next node will be a [dom.Element], so keep unwrapping that until
       /// we get the underlying text node, and finally get its text.
-      if (elementIndex < (parentNodes?.length ?? 1) - 1 && parentNodes?[elementIndex + 1] is dom.Text) {
+      if (elementIndex < (parentNodes?.length ?? 1) - 1 &&
+          parentNodes?[elementIndex + 1] is dom.Text) {
         parentAfterText = parentNodes?[elementIndex + 1].text ?? " ";
       } else if (elementIndex < (parentNodes?.length ?? 1) - 1) {
         var parentAfter = parentNodes?[elementIndex + 1];
@@ -441,6 +478,7 @@ class HtmlParser extends StatelessWidget {
         }
         parentAfterText = parentAfter?.text ?? " ";
       }
+
       /// If the text is the first element in the current tree node list, it
       /// starts with a whitespace, it isn't a line break, either the
       /// whitespace is unnecessary or it is a block element, and either it is
@@ -450,38 +488,38 @@ class HtmlParser extends StatelessWidget {
       /// We should also delete the whitespace at any point in the node list
       /// if the previous element is a <br> because that tag makes the element
       /// act like a block element.
-      if (textIndex < 1
-          && tree.text!.startsWith(' ')
-          && tree.element?.localName != "br"
-          && (!keepLeadingSpace.data
-              || BLOCK_ELEMENTS.contains(tree.element?.localName ?? ""))
-          && (elementIndex < 1
-              || (elementIndex >= 1
-                  && parentNodes?[elementIndex - 1] is dom.Text
-                  && parentNodes![elementIndex - 1].text!.endsWith(" ")))
-      ) {
+      if (textIndex < 1 &&
+          tree.text!.startsWith(' ') &&
+          tree.element?.localName != "br" &&
+          (!keepLeadingSpace.data ||
+              BLOCK_ELEMENTS.contains(tree.element?.localName ?? "")) &&
+          (elementIndex < 1 ||
+              (elementIndex >= 1 &&
+                  parentNodes?[elementIndex - 1] is dom.Text &&
+                  parentNodes![elementIndex - 1].text!.endsWith(" ")))) {
         tree.text = tree.text!.replaceFirst(' ', '');
-      } else if (textIndex >= 1
-          && tree.text!.startsWith(' ')
-          && tree.element?.nodes[textIndex - 1] is dom.Element
-          && (tree.element?.nodes[textIndex - 1] as dom.Element).localName == "br"
-      ) {
+      } else if (textIndex >= 1 &&
+          tree.text!.startsWith(' ') &&
+          tree.element?.nodes[textIndex - 1] is dom.Element &&
+          (tree.element?.nodes[textIndex - 1] as dom.Element).localName ==
+              "br") {
         tree.text = tree.text!.replaceFirst(' ', '');
       }
+
       /// If the text is the last element in the current tree node list, it isn't
       /// a line break, and the next text node starts with a whitespace,
       /// update the [Context] to signify to that next text node whether it should
       /// keep its whitespace. This is based on whether the current text ends with a
       /// whitespace.
-      if (textIndex == (tree.element?.nodes.length ?? 1) - 1
-          && tree.element?.localName != "br"
-          && parentAfterText.startsWith(' ')
-      ) {
+      if (textIndex == (tree.element?.nodes.length ?? 1) - 1 &&
+          tree.element?.localName != "br" &&
+          parentAfterText.startsWith(' ')) {
         keepLeadingSpace.data = !tree.text!.endsWith(' ');
       }
     }
 
-    tree.children.forEach((e) => _processInlineWhitespaceRecursive(e, keepLeadingSpace));
+    tree.children
+        .forEach((e) => _processInlineWhitespaceRecursive(e, keepLeadingSpace));
 
     return tree;
   }
@@ -518,14 +556,19 @@ class HtmlParser extends StatelessWidget {
     if (tree.style.listStylePosition == null) {
       tree.style.listStylePosition = ListStylePosition.OUTSIDE;
     }
-    if (tree.name == 'ol' && tree.style.listStyleType != null && tree.style.listStyleType!.type == "marker") {
+    if (tree.name == 'ol' &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "marker") {
       switch (tree.style.listStyleType!) {
         case ListStyleType.LOWER_LATIN:
         case ListStyleType.LOWER_ALPHA:
         case ListStyleType.UPPER_LATIN:
         case ListStyleType.UPPER_ALPHA:
           olStack.add(Context<String>('a'));
-          if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+          if ((tree.attributes['start'] != null
+                  ? int.tryParse(tree.attributes['start']!)
+                  : null) !=
+              null) {
             var start = int.tryParse(tree.attributes['start']!) ?? 1;
             var x = 1;
             while (x < start) {
@@ -535,14 +578,22 @@ class HtmlParser extends StatelessWidget {
           }
           break;
         default:
-          olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+          olStack.add(Context<int>((tree.attributes['start'] != null
+                  ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                  : 1) -
+              1));
           break;
       }
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null && tree.style.listStyleType!.type == "widget") {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "widget") {
       tree.style.markerContent = tree.style.listStyleType!.widget!;
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null && tree.style.listStyleType!.type == "image") {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null &&
+        tree.style.listStyleType!.type == "image") {
       tree.style.markerContent = Image.network(tree.style.listStyleType!.text);
-    } else if (tree.style.display == Display.LIST_ITEM && tree.style.listStyleType != null) {
+    } else if (tree.style.display == Display.LIST_ITEM &&
+        tree.style.listStyleType != null) {
       String marker = "";
       switch (tree.style.listStyleType!) {
         case ListStyleType.NONE:
@@ -558,7 +609,10 @@ class HtmlParser extends StatelessWidget {
           break;
         case ListStyleType.DECIMAL:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           marker = '${olStack.last.data}.';
@@ -567,7 +621,10 @@ class HtmlParser extends StatelessWidget {
         case ListStyleType.LOWER_ALPHA:
           if (olStack.isEmpty) {
             olStack.add(Context<String>('a'));
-            if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+            if ((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start']!)
+                    : null) !=
+                null) {
               var start = int.tryParse(tree.attributes['start']!) ?? 1;
               var x = 1;
               while (x < start) {
@@ -583,7 +640,10 @@ class HtmlParser extends StatelessWidget {
         case ListStyleType.UPPER_ALPHA:
           if (olStack.isEmpty) {
             olStack.add(Context<String>('a'));
-            if ((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start']!) : null) != null) {
+            if ((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start']!)
+                    : null) !=
+                null) {
               var start = int.tryParse(tree.attributes['start']!) ?? 1;
               var x = 1;
               while (x < start) {
@@ -597,18 +657,27 @@ class HtmlParser extends StatelessWidget {
           break;
         case ListStyleType.LOWER_ROMAN:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           if (olStack.last.data <= 0) {
             marker = '${olStack.last.data}.';
           } else {
-            marker = (olStack.last.data as int).toRomanNumeralString()!.toLowerCase() + ".";
+            marker = (olStack.last.data as int)
+                    .toRomanNumeralString()!
+                    .toLowerCase() +
+                ".";
           }
           break;
         case ListStyleType.UPPER_ROMAN:
           if (olStack.isEmpty) {
-            olStack.add(Context<int>((tree.attributes['start'] != null ? int.tryParse(tree.attributes['start'] ?? "") ?? 1 : 1) - 1));
+            olStack.add(Context<int>((tree.attributes['start'] != null
+                    ? int.tryParse(tree.attributes['start'] ?? "") ?? 1
+                    : 1) -
+                1));
           }
           olStack.last.data += 1;
           if (olStack.last.data <= 0) {
@@ -619,8 +688,8 @@ class HtmlParser extends StatelessWidget {
           break;
       }
       tree.style.markerContent = Text(
-          marker,
-          textAlign: TextAlign.right,
+        marker,
+        textAlign: TextAlign.right,
       );
     }
 
@@ -639,11 +708,17 @@ class HtmlParser extends StatelessWidget {
   static StyledElement _processBeforesAndAfters(StyledElement tree) {
     if (tree.style.before != null) {
       tree.children.insert(
-          0, TextContentElement(text: tree.style.before, style: tree.style.copyWith(beforeAfterNull: true, display: Display.INLINE)));
+          0,
+          TextContentElement(
+              text: tree.style.before,
+              style: tree.style
+                  .copyWith(beforeAfterNull: true, display: Display.INLINE)));
     }
     if (tree.style.after != null) {
-      tree.children
-          .add(TextContentElement(text: tree.style.after, style: tree.style.copyWith(beforeAfterNull: true, display: Display.INLINE)));
+      tree.children.add(TextContentElement(
+          text: tree.style.after,
+          style: tree.style
+              .copyWith(beforeAfterNull: true, display: Display.INLINE)));
     }
 
     tree.children.forEach(_processBeforesAndAfters);
@@ -766,13 +841,13 @@ class HtmlParser extends StatelessWidget {
     tree.children.forEach((child) {
       if (child is EmptyContentElement || child is EmptyLayoutElement) {
         toRemove.add(child);
-      } else if (child is TextContentElement
-          && (tree.name == "body" || tree.name == "ul")
-          && child.text!.replaceAll(' ', '').isEmpty) {
+      } else if (child is TextContentElement &&
+          (tree.name == "body" || tree.name == "ul") &&
+          child.text!.replaceAll(' ', '').isEmpty) {
         toRemove.add(child);
-      } else if (child is TextContentElement
-          && child.text!.isEmpty
-          && child.style.whiteSpace != WhiteSpace.PRE) {
+      } else if (child is TextContentElement &&
+          child.text!.isEmpty &&
+          child.style.whiteSpace != WhiteSpace.PRE) {
         toRemove.add(child);
       } else if (child is TextContentElement &&
           child.style.whiteSpace != WhiteSpace.PRE &&
@@ -852,7 +927,7 @@ class ContainerSpan extends StatelessWidget {
     required this.style,
     required this.newContext,
     this.shrinkWrap = false,
-  }): super(key: key);
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext _) {
@@ -897,7 +972,7 @@ class StyledText extends StatelessWidget {
     this.key,
     this.selectionControls,
     this.scrollPhysics,
-  }) : _selectable = false,
+  })  : _selectable = false,
         super(key: key);
 
   const StyledText.selectable({
@@ -908,7 +983,7 @@ class StyledText extends StatelessWidget {
     this.key,
     this.selectionControls,
     this.scrollPhysics,
-  }) : textSpan = textSpan,
+  })  : textSpan = textSpan,
         _selectable = true,
         super(key: key);
 
@@ -941,7 +1016,8 @@ class StyledText extends StatelessWidget {
   }
 
   double? consumeExpandedBlock(Display? display, RenderContext context) {
-    if ((display == Display.BLOCK || display == Display.LIST_ITEM) && !renderContext.parser.shrinkWrap) {
+    if ((display == Display.BLOCK || display == Display.LIST_ITEM) &&
+        !renderContext.parser.shrinkWrap) {
       return double.infinity;
     }
     return null;
@@ -952,7 +1028,8 @@ extension IterateLetters on String {
   String nextLetter() {
     String s = this.toLowerCase();
     if (s == "z") {
-      return String.fromCharCode(s.codeUnitAt(0) - 25) + String.fromCharCode(s.codeUnitAt(0) - 25); // AA or aa
+      return String.fromCharCode(s.codeUnitAt(0) - 25) +
+          String.fromCharCode(s.codeUnitAt(0) - 25); // AA or aa
     } else {
       var lastChar = s.substring(s.length - 1);
       var sub = s.substring(0, s.length - 1);
